+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 2.1.8 - 2020-03-31
+
+### Added
+
+- Add `useFirstIceCandidate` parameters to `NXMClientConfig`
+
+```
+NXMClientConfig *config = [[NXMClientConfig alloc] initWithApiUrl:restUrl
+                                                     websocketUrl:wsUrl
+                                                           ipsUrl:ipsUrl
+                                             useFirstIceCandidate:NO];
+```
 
 ## 2.1.5 - 2020-03-18